split logic in files, use data structures from the domain
