+++ track/render.go
@@ -2,7 +2,6 @@ package track
 
 import (
 	"fmt"
-	"math/rand"
 	"time"
 
 	"github.com/faiface/pixel"
@@ -12,48 +11,33 @@ import (
 	"golang.org/x/image/colornames"
 )
 
-func drawTrack() *imdraw.IMDraw {
+func drawTrack(track domain.Track) *imdraw.IMDraw {
 
-	rand.Seed(time.Now().UnixNano())
+	trackRender := imdraw.New(nil)
+	trackRender.Color = pixel.RGB(1, 0, 0)
 
-	points := 20
-	maxPoints := 100
-	width := int32(1024)
-	height := int32(768)
-
-	var pointsLeft = make([]domain.Position, points, maxPoints)
-	var pointsRight = make([]domain.Position, points, maxPoints)
-
-	for i := 0; i < points; i++ {
-		pointsLeft = append(pointsLeft, domain.Position{X: float32(rand.Int31n(width)), Y: float32(rand.Int31n(height))})
-		pointsRight = append(pointsRight, domain.Position{X: float32(rand.Int31n(width)), Y: float32(rand.Int31n(height))})
-	}
-
-	track := imdraw.New(nil)
-	track.Color = pixel.RGB(1, 0, 0)
-
-	for i := 0; i < len(pointsLeft)-1; i = i + 2 {
-		pointA := pointsLeft[i]
-		pointB := pointsLeft[i+1]
-		track.Push(
+	for i := 0; i < len(track.FirstSide)-1; i = i + 2 {
+		pointA := track.FirstSide[i]
+		pointB := track.FirstSide[i+1]
+		trackRender.Push(
 			pixel.V(float64(pointA.X), float64(pointA.Y)),
 			pixel.V(float64(pointB.X), float64(pointB.Y)),
 		)
-		track.Line(10)
+		trackRender.Line(10)
 	}
-	track.Color = pixel.RGB(0, 0, 1)
+	trackRender.Color = pixel.RGB(0, 0, 1)
 
-	for i := 0; i < len(pointsRight)-1; i = i + 2 {
-		pointA := pointsRight[i]
-		pointB := pointsRight[i+1]
-		track.Push(
+	for i := 0; i < len(track.SecondSide)-1; i = i + 2 {
+		pointA := track.SecondSide[i]
+		pointB := track.SecondSide[i+1]
+		trackRender.Push(
 			pixel.V(float64(pointA.X), float64(pointA.Y)),
 			pixel.V(float64(pointB.X), float64(pointB.Y)),
 		)
-		track.Line(10)
+		trackRender.Line(10)
 	}
 
-	return track
+	return trackRender
 }
 
 func run() {
@@ -74,11 +58,12 @@ func run() {
 		second = time.Tick(time.Second)
 	)
 
-	track := drawTrack()
+	track := genTrack()
+	trackRender := drawTrack(track)
 
 	for !win.Closed() {
 		win.Clear(colornames.Skyblue)
-		track.Draw(win)
+		trackRender.Draw(win)
 		win.Update()
 
 		frames++
